requirements.txt
# Packages without specific version restriction.
discord.py
aiohttp
motor
ujson
numpy
pillow
aioredis

# Packages from GitHub.
jishaku @ git+https://github.com/Gorialis/jishaku@master
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@master
discord-ext-menus-views @ git+https://github.com/oliver-ni/discord-ext-menus-views@master